refactor: add config files to the package src and a command to generate config
This allows access to the config for people installing the repo with pip to use
the command line. The config templates are in `shimmer_ssd/config_template` and
the config folder can be created with `ssd config create`.
Also add a `main.yaml` file for all config instead of `local.yaml` (local.yaml still
works).

Also now all command default to a config location in `./config`.